\name{columns}
\docType{methods}

\alias{generic,columns}
\alias{columns,MSEADb-method}
\alias{columns}

\title{A function to return the columns of MSEA that user can retrieve}
\description{This function returns the columns of MSEA that
user can retrieve.}

\usage{columns(x)}

\arguments{
\item{x}{\code{MSEADb} object such as \code{MSEA.Ath.pb.db}}
}
\value{
The columns of sqlite database
}
\author{Kozo Nishida}
\examples{
showMethods("columns")
}
