% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeMSEAPackage.R
\name{makeMSEAPackage}
\alias{makeMSEAPackage}
\title{Making MSEADb packages from corresponding table as single data frame.}
\usage{
makeMSEAPackage(
  pkgname,
  data,
  metadata,
  organism,
  version,
  maintainer,
  author,
  destDir,
  license = "Artistic-2.0"
)
}
\arguments{
\item{pkgname}{What is the desired package name.}

\item{data}{Data frame contains GENEID (e.g., 100036770), MESHID
(e.g.D000465), CATEGORY (e.g., G), SOURCEID (pubmed id), and SOURCEDB (e.g.,
gendoo)}

\item{metadata}{Data frame contains metadata of the package}

\item{organism}{The name of the organism this package represents}

\item{version}{What is the version number for this package?}

\item{maintainer}{Who is the package maintainer? (must include email to be
valid)}

\item{author}{Who is the creator of this package?}

\item{destDir}{A path where the package source should be assembled.}

\item{license}{What is the license (and it's version)}
}
\value{
A special package to load an \link{MSEADb} object.
}
\description{
\code{makeMSEAPackage} is a method that generates a package that will
load an appropriate \code{MSEADb} object that will in turn point to existing
annotation packages.
}
\details{
The purpose of this method is to create a special package that will depend
on existing annotation packages and which will load a special \code{MSEADb}
object that will allow proper dispatch of special select methods. These
methods will allow the user to easily query across multiple annotation
resources via information contained by the \code{MSEADb} object. Because the
end result will be a package that treats all the data mapped together as a
single source, the user is encouraged to take extra care to ensure that the
different packages used are from the same build etc.
}
\examples{

## makeMSEAPackage enable users to construct
## user's own custom MSEA annotation package

## this is test data which means the relationship between
## PathBank pathway IDs of Arabidopsis thaliana
## and its compound DB IDs (e.g., HMDB, CAS, etc...).
tmp <- tempdir()
ath <- system.file("extdata","MSEA.Ath.pb.db_DATA.csv",package="MSEADbi")
meta <- system.file("extdata","MSEA.Ath.pb.db_METADATA.csv",
    package="MSEADbi")
athDf <- read.csv(ath)
metaDf <- read.csv(meta)
# We need to avoid DOT from the column names (to query with the names)
names(athDf) <- gsub("\\\\.", "", names(athDf))
names(metaDf) <- gsub("\\\\.", "", names(metaDf))

makeMSEAPackage(pkgname = "MSEA.Ath.pb.db", data=athDf, metadata=metaDf,
    organism = "Arabidopsis thaliana", version = "0.99.0",
    maintainer = "Kozo Nishida <kozo.nishida@gmail.com>",
    author = "Kozo Nishida",
    destDir = tmp, license = "Artistic-2.0")

mseaPackageDir = paste(tmp, "MSEA.Ath.pb.db", sep="/")
install.packages(mseaPackageDir, repos=NULL, type="source")

}
\seealso{
\code{\link{MSEADb}}
}
\author{
Koki Tsuyuzaki
}
