\name{mseaListPathwaySubjects}
\docType{methods}

\alias{generic, mseaListPathwaySubjects}
\alias{mseaListPathwaySubjects,MSEADb-method}
\alias{mseaListPathwaySubjects}

\title{A function to return the subject of pathways used in the package}
\description{This function returns the list of pathway subjects used in the
package.}

\usage{mseaListPathwaySubjects(x)}

\arguments{
\item{x}{\code{MSEADb} object such as \code{MSEA.Ath.pb.db}}
}
\value{
The list of pathway subjects
}
\author{Kozo Nishida}
\examples{
showMethods("mseaListPathwaySubjects")
}
