\name{mseaNomenclature}
\docType{methods}

\alias{generic,mseaNomenclature}
\alias{mseaNomenclature,MSEADb-method}
\alias{mseaNomenclature}

\title{A function to return the scientific name}
\description{This function returns the scientific name of
species used in the package.}

\usage{mseaNomenclature(x)}

\arguments{
\item{x}{\code{MSEADb} object such as \code{MSEA.Ath.pb.db}}
}
\value{
The scientific name of organisms
}
\author{Kozo Nishida}
\examples{
showMethods("mseaNomenclature")
}
