\name{MSEAp.Ath.pb.db_DATA}
\alias{MSEAp.Ath.pb.db_DATA}
\docType{data}
\title{
Data to construct user's original MSEApDb package
}
\description{
Correspondance between PathBank pathway IDs of Arabidopsis thaliana and other compound databases' IDs is provided as demo data. This is to demonstrate how to construct user's original \code{MSEApDb} package. The data is based on the PathBank database.

}

\usage{data(MSEAp.Ath.pb.db_DATA)}


\examples{
data(MSEAp.Ath.pb.db_DATA)
head(MSEAp.Ath.pb.db_DATA)
}

\keyword{datasets}