\name{MSEAp.Ath.pb.db_METADATA}
\alias{MSEAp.Ath.pb.db_METADATA}
\docType{data}
\title{
Metadata to construct user's original MSEApDb package
}
\description{Meta data to construct user's custom MSEApDb}

\usage{data(MSEAp.Ath.pb.db_METADATA)}


\examples{
data(MSEAp.Ath.pb.db_METADATA)
head(MSEAp.Ath.pb.db_METADATA)
}

\keyword{datasets}