\name{mseaPackageName}
\docType{methods}

\alias{generic,mseaPackageName}
\alias{mseaPackageName,MSEADb-method}
\alias{mseaPackageName}

\title{A function to return the package name}
\description{This function returns the package name}

\usage{mseaPackageName(x)}

\arguments{
\item{x}{\code{MSEADb} object such as \code{MSEA.Ath.pb.db}}
}
\value{
The name of MSEA.XXX.pb.db package
}
\author{Kozo Nishida}
\examples{
showMethods("mseaPackageName")
}
