\name{mseapListPathwaySubjects}
\docType{methods}

\alias{generic,mseapListPathwaySubjects}
\alias{mseapListPathwaySubjects,MSEApDb-method}
\alias{mseapListPathwaySubjects}

\title{A function to return the scientific name of package}
\description{This function returns the scientific name of species used in the package.}

\usage{mseapListPathwaySubjects(x)}

\arguments{
\item{x}{\code{MSEApDb} object such as \code{MSEAp.Ath.pb.db}}
}
\author{Kozo Nishida}
\examples{
# library("MSEAp.Ath.pb.db")
# mseapListPathwaySubjects(MSEAp.Ath.pb.db)
}