\name{mseapPackageName}
\docType{methods}

\alias{generic,mseapPackageName}
\alias{mseapPackageName,MSEApDb-method}
\alias{mseapPackageName}

\title{A function to return the name of package}
\description{This function returns the name of package}

\usage{mseapPackageName(x)}

\arguments{
\item{x}{\code{MSEApDb} object such as \code{MSEAp.Ath.pb.db}}
}
\author{Kozo Nishida}
\examples{
# library("MSEAp.Ath.pb.db")
# mseapPackageName(MSEAp.Ath.pb.db)
}