\name{mseaVersion}
\docType{methods}

\alias{generic,mseaVersion}
\alias{mseaVersion,MSEADb-method}
\alias{mseaVersion}

\title{A function to return the MSEA data source version of package}
\description{This function returns the version of MSEA data source used in the
package.}

\usage{mseaVersion(x)}

\arguments{
\item{x}{\code{MSEADb} object such as \code{MSEA.Ath.pb.db}}
}
\value{
The version of MSEA database
}
\author{Kozo Nishida}
\examples{
showMethods("mseaVersion")
}
