\name{species}
\docType{methods}

\alias{generic,species}
\alias{species,MSEADb-method}
\alias{species}

\title{A function to return the name of organism}
\description{This function returns the name of organism.}

\usage{species(object)}

\arguments{
\item{object}{\code{MSEADb} object such as \code{MSEA.Ath.pb.db}}
}
\value{
The common name of organisms
}
\author{Kozo Nishida}
\examples{
showMethods("species")
}
